extract sign in logic into authProvider
