+++ app/authProvider.tsx
@@ -8,15 +8,21 @@ import {
   useMemo,
   useCallback,
 } from "react";
-import { User } from "firebase/auth";
+import {
+  FacebookAuthProvider,
+  GoogleAuthProvider,
+  User,
+  signInWithPopup,
+} from "firebase/auth";
 import { useRouter } from "next/navigation";
 
 import { onAuthStateChanged } from "@/lib/auth";
 import { auth } from "@/server/initFirebase";
+import { mapUser } from "@/server/mapper/userMapper";
 
 export interface AuthContext {
   user: User | null;
-  signIn: () => void;
+  signIn: (providerName: "Google" | "Facebook") => void;
   signOut: () => void;
 }
 
@@ -27,7 +33,32 @@ export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
   const [user, setUser] = useState<User | null>(null);
   const router = useRouter();
 
-  const signIn = useCallback(async () => {
+  const signIn = useCallback(async (providerName: "Google" | "Facebook") => {
+    let credential, provider;
+
+    try {
+      switch (providerName) {
+        case "Google":
+          provider = new GoogleAuthProvider();
+          credential = await signInWithPopup(auth, provider);
+          break;
+        case "Facebook":
+          provider = new FacebookAuthProvider();
+          credential = await signInWithPopup(auth, provider);
+          break;
+      }
+      const user = mapUser(credential, 1);
+
+      await fetch("/api/sign-in/", {
+        method: "PATCH",
+        body: JSON.stringify({
+          ...user,
+          guid: credential.user.uid,
+        }),
+      });
+    } catch (error) {
+      console.error(`Error signing in with ${providerName}.`, error);
+    }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
@@ -45,6 +76,7 @@ export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
     } catch (error) {
       console.error("Error signing out.", error);
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   useEffect(() => {
